Move pipeline_state and pipeline_graph to agent/, add bridge/ shims for backward compat
